Add logic to mount smb
