+++ apps/vs-code-designer/src/app/tree/subscriptionTree/SubscriptionTreeItem.ts
@@ -6,8 +6,8 @@ import { LogicAppResolver } from '../../../LogicAppResolver';
 import { projectLanguageSetting, webProvider, workflowappRuntime, storageProvider, insightsProvider } from '../../../constants';
 import { ext } from '../../../extensionVariables';
 import { localize } from '../../../localize';
-import { ConnectEnvironmentStep } from '../../commands/createLogicApp/createLogicAppSteps/HybridLogicAppsSteps/ConnectEnvironmentStep';
-import { ContainerAppCreateStep } from '../../commands/createLogicApp/createLogicAppSteps/HybridLogicAppsSteps/ContainerAppCreateStep';
+// import { ConnectEnvironmentStep } from '../../commands/createLogicApp/createLogicAppSteps/HybridLogicAppsSteps/ConnectEnvironmentStep';
+// import { ContainerAppCreateStep } from '../../commands/createLogicApp/createLogicAppSteps/HybridLogicAppsSteps/ContainerAppCreateStep';
 import { LogicAppCreateStep } from '../../commands/createLogicApp/createLogicAppSteps/LogicAppCreateStep';
 import { LogicAppHostingPlanStep } from '../../commands/createLogicApp/createLogicAppSteps/LogicAppHostingPlanStep';
 import { AzureStorageAccountStep } from '../../commands/deploy/storageAccountSteps/AzureStorageAccountStep';
@@ -152,8 +152,8 @@ export class SubscriptionTreeItem extends SubscriptionTreeItemBase {
 
     executeSteps.push(new VerifyProvidersStep([webProvider, storageProvider, insightsProvider]));
     executeSteps.push(new LogicAppCreateStep());
-    executeSteps.push(new ConnectEnvironmentStep());
-    executeSteps.push(new ContainerAppCreateStep());
+    // executeSteps.push(new ConnectEnvironmentStep());
+    // executeSteps.push(new ContainerAppCreateStep());
 
     const title: string = localize('functionAppCreatingTitle', 'Create new Logic App (Standard) in Azure');
     const wizard: AzureWizard<IAppServiceWizardContext> = new AzureWizard(wizardContext, { promptSteps, executeSteps, title });
@@ -220,6 +220,7 @@ export class SubscriptionTreeItem extends SubscriptionTreeItemBase {
     const slotTreeItem = new SlotTreeItem(subscription, resolved);
     slotTreeItem.customLocation = wizardContext.customLocation;
     slotTreeItem.fileShare = wizardContext.fileShare;
+    slotTreeItem.isHybridLogicApp = wizardContext.useHybrid;
     return slotTreeItem;
   }
